Update project configuration and files
