+++ package.json
@@ -8,7 +8,12 @@
     "build": "vite build"
   },
   "devDependencies": {
+    "@types/react": "^18.2.69",
+    "@types/react-dom": "^18.2.22",
+    "@types/three": "^0.162.0",
     "@vitejs/plugin-react": "4.1",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.4.3",
     "vite": "^4.5.0",
     "vite-plugin-glsl": "^1.1.2"
   },
