+++ src/index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script type="module" src="./index.jsx"></script>
+    <script type="module" src="./index.tsx"></script>
 </body>
 </html>
