+++ tailwind.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    content: ["./src/**/*.{html,js,jsx,tsx}"],
+    theme: {
+      extend: {},
+    },
+    plugins: [],
+  };
