+++ tsconfig.node.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "composite": true,
+        "module": "esnext",
+        "moduleResolution": "node"
+      },
+      "include": ["vite.config.ts"]
+}
