--- vite.config.js
@@ -1,42 +1,0 @@
-import glsl from 'vite-plugin-glsl'
-import react from '@vitejs/plugin-react'
-import { transformWithEsbuild } from 'vite'
-
-
-export default {
-    root: 'src/',
-    publicDir: '../static/',
-    base: './',
-    server:
-    {
-        host: true, // Open to local network and display URL
-        open: !('SANDBOX_URL' in process.env || 'CODESANDBOX_HOST' in process.env) // Open if it's not a CodeSandbox
-    },
-    build:
-    {
-        outDir: '../dist', // Output in the dist/ folder
-        emptyOutDir: true, // Empty the folder first
-        sourcemap: true // Add sourcemap
-    },
-    plugins:
-    [
-        glsl(),
-        // React support
-        react(),
-
-        // .js file support as if it was JSX
-        {
-            name: 'load+transform-js-files-as-jsx',
-            async transform(code, id)
-            {
-                if (!id.match(/src\/.*\.js$/))
-                    return null
-
-                return transformWithEsbuild(code, id, {
-                    loader: 'jsx',
-                    jsx: 'automatic',
-                });
-            },
-        },
-    ]
-}
