+++ vite.config.ts
@@ -1,0 +1,23 @@
+import glsl from 'vite-plugin-glsl'
+import react from '@vitejs/plugin-react'
+
+export default {
+    root: 'src/',
+    publicDir: '../static/',
+    base: './',
+
+    server: {
+        host: true, // Open to local network and display URL
+    },
+    build:
+    {
+        outDir: '../dist', // Output in the dist/ folder
+        emptyOutDir: true, // Empty the folder first
+        sourcemap: true // Add sourcemap
+    },
+    plugins:
+    [
+        glsl(),
+        react(),
+    ]
+}
